fix(content-releases): fix review comments
